Refactor code common to GPU profilers into separate CUs
