+++ onnxruntime/core/providers/cuda/cupti_manager.h
@@ -8,8 +8,9 @@
 
 #include <cupti.h>
 
+#include "core/common/gpu_profiler_common.h"
 #include "core/common/inlined_containers.h"
-#include "core/common/profiler_common.h"
+
 
 
 namespace onnxruntime {
@@ -23,12 +24,12 @@ public:
     ORT_DISALLOW_COPY_ASSIGNMENT_AND_MOVE(CUPTIManager);
     ~CUPTIManager();
     static CUPTIManager& GetInstance();
-    void StartLogging() override;
 
 protected:
   bool PushUniqueCorrelation(uint64_t unique_cid) override;
   void PopUniqueCorrelation(uint64_t& popped_unique_cid) override;
-  void StopLogging() override;
+  bool OnStartLogging() override;
+  void OnStopLogging() override;
   void ProcessActivityBuffers(const std::vector<ProfilerActivityBuffer>& buffers,
                               const TimePoint& start_time) override;
   void FlushActivities() override;
@@ -37,6 +38,12 @@ private:
     static constexpr size_t kActivityBufferSize = 32 * 1024;
     static constexpr size_t kActivityBufferAlignSize = 8;
 
+    // TODO: Is this even needed? malloc() is required to return
+    // a memory block that meets the alignment requirements for _any_ data type.
+    // On any platform that supports an 8-byte datatype (double? long long?)
+    // this means that malloc() already returns memory aligned at
+    // _at least_ 8 byte boundaries, rendering this additional alignment
+    // redundant?
     static constexpr uint8_t* AlignBuffer(uint8_t* buffer, int align) {
         return (((uintptr_t)(buffer) & ((align)-1))
                 ? ((buffer) + (align) - ((uintptr_t)(buffer) & ((align)-1)))
